STYLE: Remove `static_cast<long>` from AdvancedImageToImageMetric
The cast is unnecessary. `Math::Round<int64_t>(x)` can be assigned to `index[j]` without casting, as `index[j]` refers to a 64-bit signed integer.

+++ Common/CostFunctions/itkAdvancedImageToImageMetric.hxx
@@ -443,7 +443,7 @@ AdvancedImageToImageMetric<TFixedImage, TMovingImage>::EvaluateMovingImageValueA
         MovingImageIndexType index;
         for (unsigned int j = 0; j < MovingImageDimension; ++j)
         {
-          index[j] = static_cast<long>(Math::Round<int64_t>(cindex[j]));
+          index[j] = Math::Round<int64_t>(cindex[j]);
         }
         (*gradient) = Superclass::m_GradientImage->GetPixel(index);
       }
